Create a new man page "freeciv" for branch S3_1, fixes ticket 42138
Create a man page, with source in doc/man/freeciv.6.in . Content and
structure are both adapted from doc/man/freeciv-{client,server}.6.in .
This page gives an overview of the Freeciv game, including its client
and server parts and its flexible rulesets and parameters. It points the
reader to the other man pages and websites for detailed information.

The problem which this page solves is that a user new to Freeciv, unsure
where to start, can type the obvious thing, "man freeciv", and get a
useful overview instead of finding nothing.

Make changes to ancillary files to incorporate the new source file.
All these changes are parallel to corresponding entries for similar man
pages.
configure.ac: include new file in AC_CONFIG_FILES([...])
doc/man/Makefile.am: include new file in MAN_PAGES
doc/man/.gitignore: ignore derived file freeciv.6

In the freeciv(6) man page source, in the "See Also" section, there is
a complete list of the other man pages provided with Freeciv. The list
of man pages is different in the S3_1 branch than in master. Make the
"See Also" section match the man pages in S3_1. Set the man page date
to today.

See osdn #42138.

Signed-off-by: Jim DeLaHunt <from+project_by_git@jdlh.com>