Change man freeciv page's See Also list to match S3_0 branch.
In the freeciv(6) man page source, in the "See Also" section, there is
a complete list of the other man pages provided with Freeciv. The list
of man pages is different in the S3_0 branch than in master. Make the
"See Also" section match the man pages in S3_0. Set the man page date
to today.

See osdn #42138.

Signed-off-by: Jim DeLaHunt <from+project_by_git@jdlh.com>